Install notes for Meld 3.20.3 on Win, fixes #187
As discussed [here](https://gitlab.gnome.org/GNOME/meld/-/issues/559), Meld 3.20.3 switched to local installation as default. This changed the default install location and also led to some errors where the installer tries to write to the Registry but without elevated privileges. The developers indicate that they might revert to system-wide install in later versions which means that this version might be the only one to have this bug. NB. Future versions will have a (non-default) option to do user only installs that might be suggested in the instructions.